+++ beam/beam/doctype/handling_unit/__init__.py
@@ -1,3 +1,6 @@
+# Copyright (c) 2025, AgriTheory and contributors
+# For license information, please see license.txt
+
 import frappe
 from pypika import Order