+++ beam/beam/doctype/handling_unit/handling_unit.js
@@ -1,3 +1,6 @@
+// Copyright (c) 2025, AgriTheory and contributors
+// For license information, please see license.txt
+
 frappe.ui.form.on('Handling Unit', {
 	refresh: frm => {},
 })