+++ beam/docs/generate_matrix.py
@@ -1,4 +1,6 @@
-import textwrap
+# Copyright (c) 2025, AgriTheory and contributors
+# For license information, please see license.txt
+
 from pathlib import Path